requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
pydantic==2.11.3
pydantic-settings==2.8.1
tenacity==9.1.2
pytest==8.3.5
pytest-asyncio==0.26.0
mcp==1.9.4
google-genai==1.13.0